+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for NeoFS Node
 - Support of `GT`, `GE`, `LT` and `LE` numeric comparison operators in CLI (#2733)
 - SN eACL processing of NULL and numeric operators (#2742)
 - CLI now allows to create and print eACL with numeric filters (#2742)
+- gRPC connection limits per endpoint (#1240) 
 
 ### Fixed
 - Access to `PUT` objects no longer grants `DELETE` rights (#2261)